+++ src/components/DeleteHintList.tsx
@@ -10,7 +10,7 @@ import {
   VStack,
 } from "@chakra-ui/react";
 
-import { avatarList } from "@/data/AvatarList";
+import { avatarList } from "@/data/data";
 
 import { Hint } from "@/types/type";
