+++ src/components/game/Answer.tsx
@@ -17,7 +17,7 @@ import {
 import { CustomInput } from "@/components/common/CustomInput";
 import { VSpacer } from "@/components/common/Spacer";
 
-import { BASE_URL } from "@/data/BaseUrl";
+import { BASE_URL } from "@/data/data";
 
 import { RecoilRoom } from "@/store/Recoil";
