+++ src/components/game/DeleteHintOtherMasterUI.tsx
@@ -16,7 +16,7 @@ import {
 
 import { VSpacer } from "@/components/common/Spacer";
 
-import { avatarList } from "@/data/AvatarList";
+import { avatarList } from "@/data/data";
 
 import { RecoilRoom } from "@/store/Recoil";
