+++ src/components/game/HowToDecideTheme.tsx
@@ -16,7 +16,7 @@ import {
 
 import { VSpacer } from "@/components/common/Spacer";
 
-import { BASE_URL } from "@/data/BaseUrl";
+import { BASE_URL } from "@/data/data";
 
 import { RecoilRoom } from "@/store/Recoil";
