+++ src/components/game/InputHint.tsx
@@ -17,7 +17,7 @@ import { CustomTitleText } from "@/components/common/CustomTitleText";
 import { VSpacer } from "@/components/common/Spacer";
 import { Wait } from "@/components/game/Wait";
 
-import { BASE_URL } from "@/data/BaseUrl";
+import { BASE_URL } from "@/data/data";
 
 import { RecoilPlayer, RecoilRoom } from "@/store/Recoil";
