--- src/data/BaseUrl.tsx
@@ -1,1 +1,0 @@
-export const BASE_URL = "http://localhost:1323/";
